requirements.txt
jax>=0.4.6
jaxlib>=0.4.6
gymnax
evosax
distrax
optax
flax
numpy
brax
dm-haiku==0.0.9
hydra-core==1.3.2
